Move heartbeat timeout management to DockerService
